% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCD_main.R
\name{ABCD}
\alias{ABCD}
\title{Main function implementing Adaptive Block-based Change-point Detection (ABCD)}
\usage{
ABCD(
  y,
  P_grid = NULL,
  k = NULL,
  B = 500,
  n0 = NULL,
  n1 = NULL,
  perm.p = FALSE,
  block_list_of_lists = NULL
)
}
\arguments{
\item{y}{An n by d time series to be analyzed}

\item{P_grid}{A vector of numbers of blocks associated with each blocking structure}

\item{k}{The value k used for constructing the k-MSTs within blocks}

\item{B}{The number of permutations used to generate a permutation p-value}

\item{n0}{First observation which a test statistic is computed for}

\item{n1}{Last observation which a test statistic is computed for}

\item{perm.p}{A T/F statement on whether a permutation p-value should be computed}

\item{block_list_of_lists}{A list of lists each indicating dimension indices
for utilizing more complicated blocking structures where blocks may overlap; overrides P_grid if used}
}
\value{
A list containing:
\item{tauhat_max}{Final ABCD change-point estimate}
\item{test_stat}{Final ABCD test statistic}
\item{p_val}{The associated permutation p-value, if computed}
\item{tauhat_inds}{A length(P_grid) by 2 matrix of which block obtained the maximum statistic in each blocking structure}
\item{tauhat_cols}{A length(P_grid) by 2 matrix of the time of maximum scan statistic within each blocking structure}
\item{tauhat_maxes}{A vector of the maximum scan statistic value within each blocking structure}
\item{boot_stats}{A vector of B-1 bootstrap statistics}
}
\description{
This function implements the change-point method ABCD from the corresponding paper
on a given time series, using a range of blocking structures to test for spatially
clustered change-points at multiple scales
}
\examples{
# Usage of ABCD with automatic blocking and default settings
test <- matrix(NA, nrow = 501, ncol = 200)
for(i in 1:450){
   test[i,] <- rnorm(200)
}
for(i in 451:501){
  test[i,] <- c(rnorm(120), rnorm(80, 1/5))
}
ABCD(t(test), perm.p=TRUE, B = 500)

# Usage of ABCD with specific blocking, choice of k, B
test <- matrix(NA, nrow = 501, ncol = 200)
for(i in 1:450){
   test[i,] <- rnorm(200)
}
for(i in 451:501){
  test[i,] <- c(rnorm(120), rnorm(80, sd = 1.1))
}
ABCD(t(test), k = 40, P_grid = c(1, 4, 9, 21, 30), B = 200, perm.p=TRUE)

}
