% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ABCD_img_main.R
\name{ABCD_img}
\alias{ABCD_img}
\title{Main function implementing Adaptive Block-based Change-point Detection (ABCD)}
\usage{
ABCD_img(
  y,
  P_1_grid = NULL,
  P_2_grid = NULL,
  k = NULL,
  B = 500,
  n0 = NULL,
  n1 = NULL,
  perm.p = F
)
}
\arguments{
\item{y}{An d_1 by d_2 by n array to be analyzed}

\item{P_1_grid}{A vector of numbers of row divisions associated with each blocking structure}

\item{P_2_grid}{A vector of numbers of column divisions associated with each blocking structure}

\item{k}{The value k used for constructing the k-MSTs within blocks}

\item{B}{The number of permutations used to generate a permutation p-value}

\item{n0}{First observation which a test statistic is computed for}

\item{n1}{Last observation which a test statistic is computed for}

\item{perm.p}{A T/F statement on whether a permutation p-value should be computed
for utilizing more complicated blocking structures where blocks may overlap; overrides P_grid if used}
}
\value{
A list containing:
\item{tauhat_max}{Final ABCD change-point estimate}
\item{test_stat}{Final ABCD test statistic}
\item{p_val}{The associated permutation p-value, if computed}
\item{tauhat_inds}{A length(P_grid) by 2 matrix of which block obtained the maximum statistic in each blocking structure}
\item{tauhat_cols}{A length(P_grid) by 2 matrix of the time of maximum scan statistic within each blocking structure}
\item{boot_stats}{A vector of B-1 bootstrap statistics}
\item{max_stats_at_change}{A vector of the maximum two-sample test statistic at tauhat_max for each blocking structure. With max_blocks_at_change, can be used to find a locational estimate of a potential change-point detected at tauhat_max.}
\item{max_blocks_at_change}{A vector of the block with the maximum two-sample test statistic at tauhat_max for each blocking structure}
}
\description{
This function implements the change-point method ABCD from the corresponding paper
on a given time series, using a range of blocking structures to test for spatially
clustered change-points at multiple scales
}
\examples{
# Usage of ABCD_img on a scale with automatic blocking and default settings
test <- array(data = NA, dim = c(60,70, 200))
for(i in 1:60){
 for(j in 1:70){
 test[i,j,] <- rnorm(200)
}
}
top_left <- c(sample(1:51, size = 1), sample(1:61, size = 1))
for(i in top_left[1]:(top_left[1]+9)){
for(j in top_left[2]:(top_left[2]+9)){
  test[i,j,121:200] <- rnorm(80, sd = 1.15)
}
}

ABCD_img(test, B = 100, perm.p=TRUE)

}
