% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking_fns.R
\name{breaks_1d}
\alias{breaks_1d}
\title{Generate block divisions for high-dimensional data}
\usage{
breaks_1d(d, P)
}
\arguments{
\item{d}{Dimension of time series to be analyzed}

\item{P}{Number of blocks to divide time series into}
}
\value{
A vector of dimension indices used to divide time series into blocks
}
\description{
This function computes divisions between blocks
in a single blocking structure for high-dimensional data
}
