% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/blocking_fns.R
\name{breaks_img}
\alias{breaks_img}
\title{Generate block divisions for image data}
\usage{
breaks_img(side_length_1, side_length_2, P_1, P_2)
}
\arguments{
\item{side_length_1}{Number of rows of image dimensions}

\item{side_length_2}{Number of columns of image dimensions}

\item{P_1}{Number of row divisions}

\item{P_2}{Number of column divisions}
}
\value{
A list of two vectors of row/column indices used to divide the image series into blocks
\item{breaks_1}{A vector of row indices used to divide the image series into blocks}
\item{breaks_2}{A vector of column indices used to divide the image series into blocks}
}
\description{
This function computes divisions between blocks
in a single blocking structure for image data
}
