% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helper_fns.R
\name{Ebynode_gen}
\alias{Ebynode_gen}
\title{Generates Ebynode object}
\usage{
Ebynode_gen(n, E)
}
\arguments{
\item{n}{Number of observations in the time series being analyzed}

\item{E}{Edge list of a similarity graph (i.e. can take kmst(y, k=k) as input)}
}
\value{
A list of n vectors, each including edges for each node
}
\description{
This function constructs a list of vectors of edges for each node
}
