% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/auto_blocking.R
\name{generate_P_vec}
\alias{generate_P_vec}
\title{Generate automatic blocking structures}
\usage{
generate_P_vec(
  first_P = 1,
  last_P = NULL,
  S = 5,
  d = NULL,
  extra_prop = 0.25,
  search_range = 3,
  poly_order = 1.5,
  target_min_block_size = 20
)
}
\arguments{
\item{first_P}{Number of blocks in first blocking structure (with largest blocks)}

\item{last_P}{Number of blocks in last blocking structure (with smallest blocks)}

\item{S}{Total number of blocking structures}

\item{d}{Dimension of analyzed data}

\item{extra_prop}{The maximum allowable proportion of extra dimensions in last block of a given blocking structure}

\item{search_range}{Defines how many integers are checked to be a number of blocks in a blocking structure}

\item{poly_order}{Order of polynomial used to generate initial vector of values to search around for numbers of blocks.}

\item{target_min_block_size}{Targeted size of block in P_S, used if last_P unspecified}
}
\value{
A vector of integers defining the number of blocks in each blocking structure
}
\description{
This function implements an method for automatically generating
blocking structures for detecting clustered changes at multiple scales
}
