% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmst_fns.R
\name{getdis}
\alias{getdis}
\title{Compute pairwise L2 Distance}
\usage{
getdis(y)
}
\arguments{
\item{y}{Time series of dimension n by d}
}
\value{
An n by n matrix of pairwise L2 distances between observations
}
\description{
This function computes the L2 distance between all pairs of
observations in a given time series of length n
}
