% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmst_fns.R
\name{kmst}
\alias{kmst}
\title{Construct k-MST on time series}
\usage{
kmst(y = NULL, dis = NULL, k = 1)
}
\arguments{
\item{y}{Time series of dimension n by d}

\item{dis}{n by n matrix of pairwise L2 distances between observations}

\item{k}{Controls number of edges k*(n-1) of k-MST similarity graph}
}
\value{
An n by 2 matrix of observations pairs connected by an edge
}
\description{
This function computes the L2 distance between all pairs of observations
in a given time series of length n, output as an n by n matrix
}
