% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helper_fns.R
\name{perm_func}
\alias{perm_func}
\title{Helper function for Permuting observations}
\usage{
perm_func(B, n)
}
\arguments{
\item{B}{Number of permutations}

\item{n}{Number of observations in the time series being analyzed}
}
\value{
A list containing:
\item{permmatches}{A list of B vectors of permutations of 1:n}
\item{perms}{A list of B sets of vectors matching pre-permuted indices to a permutation}
}
\description{
This function helps to permute the labels of observations for running tests to
generate the permutation null distribution for computing a p-value
}
