% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/other_helper_fns.R
\name{stat_row}
\alias{stat_row}
\title{Generates a vector of graph-based scan statistics for a single block}
\usage{
stat_row(Ebynode, edge_info, n, n0, n1)
}
\arguments{
\item{Ebynode}{A list of length n of vectors of edges for a given node}

\item{edge_info}{A list of lists of links for each node in the similarity graph}

\item{n}{Number of observations in the time series being analyzed}

\item{n0}{First observation which a test statistic is computed for}

\item{n1}{Last observation which a test statistic is computed for}
}
\value{
A vector of length n1-n0 of graph-based max-type scan statistics
}
\description{
This function uses the similarity graph computed for a specific block
of dimensions to generate a vector of graph-based scan statistics
for all times from t = n0 to t = n1
}
